refactor: SideBar
