+++ frontend/src/modules/Dashboard/Pages/DashboardPage.tsx
@@ -1,9 +1,11 @@
 import {  Container } from '@mui/material';
 import { Dashboard } from '../components/Dashboard/Dashboard';
+import SideBar from '@/modules/core/components/SideBar/Sidebar';
 
 const DashboardPage = () => {
   return (
-    <Container maxWidth="xl" sx={(theme)=>({background: theme.palette.secondary.main})}>
+    <Container maxWidth="xl" sx={(theme)=>({display:"flex",background: theme.palette.secondary.main})}>
+        <SideBar/>
         <Dashboard />
     </Container>
   );
