+++ frontend/src/modules/core/components/ListNav/ListNav.tsx
@@ -8,7 +8,7 @@ import {
 import { NavLink } from "react-router-dom";
 import DashboardIcon from "@mui/icons-material/Dashboard";
 import type { DashBoardIcons } from "@/types/dashboard-icons";
-import { listItemButton, listItemIcon } from "./DashboardLayout.styles";
+import { listItemButton, listItemIcon } from "../SideBar/SideBar.styles";
 import CurrencyExchangeIcon from "@mui/icons-material/CurrencyExchange";
 
 interface Props {
