+++ frontend/src/modules/core/components/SideBar/SideBar.styles.ts
